requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# mtab/interface/__init__.py: 2
# mtab/interface/routes.py: 4
# mtab/test_1.py: 2
Flask == 1.1.2

# mtab/interface/form.py: 2
Flask_WTF == 0.15.1

# mtab/api/annotator/m_preprocess.py: 6
Markdown == 3.3.4

# mtab/interface/__init__.py: 3,4,5
Pygments == 2.8.1

# mtab/api/lookup/m_wikidata_sparql.py: 3
SPARQLWrapper == 1.8.5

# mtab/interface/form.py: 3,12
WTForms == 2.3.3

# mtab/interface/routes.py: 10
Werkzeug == 1.0.1

# mtab/api/lookup/m_wikidata.py: 5
beautifulsoup4 == 4.9.3

# mtab/tests/test_io.py: 18
brotlipy == 0.7.0

# mtab/api/m_f.py: 59
# mtab/api/utilities/m_io.py: 3,24
charamel == 1.0.0

# mtab/api/utilities/m_io.py: 27
chardet == 3.0.4

# mtab/api/utilities/m_io.py: 28
clevercsv == 0.6.8

# mtab/api/lookup/m_entity_bm25.py: 5
elasticsearch == 7.9.1

# mtab/api/lookup/m_google.py: 1
# mtab/api/lookup/m_wikidata.py: 2
# mtab/api/lookup/m_wikidata_hist.py: 11
f == 0.0.1

# mtab/api/utilities/m_lang_detector.py: 1
fasttext == 0.9.2

# mtab/api/resources/m_fdb.py: 5,7
fdb == 2.0.2

# mtab/test_1.py: 4
flask-rest-paginate == 1.1.3

# mtab/test_1.py: 3
flask-restful == 0.3.9

# mtab/test_1.py: 3
flask-restful-theodofork == 0.3.7

# mtab/interface/routes.py: 16
flask_paginate == 0.8.1

# mtab/api/utilities/m_utils.py: 14
ftfy == 6.0.3

# mtab/api/lookup/m_google.py: 3
google-api-python-client == 2.33.0

# mtab/api/lookup/m_google.py: 3
google-api-python-client-uritemplate == 1.4.2

# mtab/api/lookup/m_google.py: 44
httpsproxy-urllib2 == 1.0

# mtab/api/utilities/m_sim.py: 1
jellyfish == 0.8.2

# mtab/api/resources/m_db_fuzzy.py: 13
# mtab/api/resources/m_db_item.py: 8
# mtab/api/resources/m_db_rocks.py: 13
# mtab/api/resources/m_parser_wikipedia.py: 9
# mtab/tests/test_io.py: 8
# mtab/tests/test_lmdb.py: 1
lmdb == 1.2.1

# mtab/api/resources/m_db_fuzzy.py: 16
# mtab/api/resources/m_db_item.py: 11
# mtab/api/resources/m_db_rocks.py: 16
# mtab/tests/test_io.py: 26
lz4 == 3.1.3

# mtab/interface/routes.py: 15
# mtab/test_1.py: 5
marshmallow == 3.10.0

# mtab/interface/__init__.py: 1
mistune == 0.8.4

# mtab/api/resources/m_db_fuzzy.py: 14
# mtab/api/resources/m_db_item.py: 9
# mtab/api/resources/m_db_rocks.py: 7
# mtab/tests/test_io.py: 9
msgpack == 1.0.2

# mtab/api/resources/m_wikigraph.py: 16
networkx == 2.5

# mtab/api/annotator/m_semantic.py: 4
# mtab/api/annotator/m_structure.py: 2
# mtab/api/lookup/m_bm25.py: 1
# mtab/api/lookup/m_entity_search.py: 4
# mtab/api/resources/m_db_fuzzy.py: 15
# mtab/api/resources/m_db_item.py: 10
# mtab/api/resources/m_db_rocks.py: 8
# mtab/api/resources/m_item.py: 10
# mtab/api/resources/m_wikigraph.py: 13
# mtab/api/utilities/m_io.py: 29
# mtab/api/utilities/m_iw.py: 5
# mtab/api/utilities/m_utils.py: 15
# mtab/test.py: 217
# mtab/tests/test_io.py: 10
numpy == 1.19.1

# mtab/tests/test_io.py: 11
orjson == 3.6.1

# mtab/api/annotator/m_preprocess.py: 7
# mtab/api/semtab/m_entity_search.py: 282
pandas == 1.1.3

# mtab/api/annotator/m_preprocess.py: 8
# mtab/api/utilities/m_io.py: 30
petl == 1.7.2

# mtab/api/utilities/m_utils.py: 17
python_dateutil == 2.8.1

# mtab/api/resources/m_db_rocks.py: 24
# mtab/tests/test_io.py: 14
python_rocksdb == 0.7.0

# mtab/api/resources/m_item_labels.py: 7
# mtab/api/utilities/m_lang_detector.py: 3
python_slugify == 5.0.2

# mtab/tests/test_io.py: 15
python_snappy == 0.6.0

# mtab/api/utilities/m_sim.py: 2
rapidfuzz == 0.14.2

# mtab/api/resources/m_parser_dbpedia.py: 10,11
# mtab/api/resources/m_parser_dbpedia_v2.py: 6,7
# mtab/api/resources/m_parser_wikidata.py: 10
rdflib == 5.0.0

# mtab/api/annotator/m_duck.py: 1
# mtab/api/annotator/m_tabano.py: 8,9,10
# mtab/api/lookup/m_mtabes.py: 1,3,4
# mtab/api/lookup/m_wikidata.py: 1
# mtab/api/lookup/m_wikidata_hist.py: 3
# mtab/api/lookup/m_wikipedia.py: 1
# mtab/api/resources/m_mediawiki.py: 14
# mtab/run_semtab.py: 8,9,10
requests == 2.24.0

# mtab/api/lookup/m_bm25.py: 3,4,5
scikit_learn == 0.22.1

# mtab/api/lookup/m_bm25.py: 2
# mtab/api/resources/m_item.py: 12,13,14,15,29
# mtab/api/resources/m_wikigraph.py: 3,17,18,19,20
# mtab/api/utilities/m_utils.py: 18
scipy == 1.4.1

# mtab/api/utilities/m_spell.py: 815
setuptools == 59.6.0

# mtab/api/resources/m_parser_wikipedia.py: 10
six == 1.15.0

# mtab/api/annotator/m_spacy.py: 3
spacy == 3.1.0

# mtab/api/lookup/m_entity_fuzzy.py: 17,18
# mtab/api/utilities/m_spell.py: 814
symspellpy == 6.7.6

# mtab/api/utilities/m_io.py: 31
texttable == 1.6.3

# mtab/api/annotator/m_semantic.py: 16
# mtab/api/annotator/m_tabano.py: 11
# mtab/api/annotator/main.py: 3
# mtab/api/lookup/m_entity_bm25.py: 6
# mtab/api/lookup/m_entity_fuzzy.py: 19
# mtab/api/lookup/m_wikidata.py: 7
# mtab/api/lookup/m_wikidata_hist.py: 10
# mtab/api/lookup/m_wikidata_sparql.py: 7
# mtab/api/lookup/m_wikipedia.py: 11
# mtab/api/resources/m_db_fuzzy.py: 17
# mtab/api/resources/m_db_item.py: 12
# mtab/api/resources/m_db_rocks.py: 14
# mtab/api/resources/m_item.py: 30
# mtab/api/resources/m_item_labels.py: 9
# mtab/api/resources/m_mapping_id.py: 6
# mtab/api/resources/m_mapping_id_v2.py: 4
# mtab/api/resources/m_parser_dbpedia.py: 12
# mtab/api/resources/m_parser_dbpedia_v2.py: 8
# mtab/api/resources/m_parser_wikidata.py: 11
# mtab/api/resources/m_parser_wikipedia.py: 12
# mtab/api/resources/m_wikigraph.py: 12
# mtab/api/semtab/m_entity_search.py: 12
# mtab/api/utilities/m_io.py: 35
# mtab/run_semtab.py: 11
# mtab/test.py: 12
# mtab/tests/test_io.py: 17
tqdm == 4.54.1

# mtab/api/resources/m_parser_wikidata.py: 9
ujson == 4.0.1

# mtab/api/resources/m_parser_wikipedia.py: 11
# mtab/api/utilities/m_utils.py: 16
wikitextparser == 0.47.5
